upgrade libz from version 1.2.3 to 1.2.5: - Version 1.2.5 fixes bugs in gzseek() and gzeof() that were present in   version 1.2.4 (March 2010). All users are encouraged to upgrade immediately. - Version 1.2.4 has many changes over 1.2.3, including these improvements:   - Fixed bugs in adler32_combine(), compressBound(), and deflateBound()   - Wholesale replacement of gz* functions with faster versions   - As part of that, added gzbuffer(), gzoffset(), gzclose_r(), and gzclose_w()     functions   - Faster Z_HUFFMAN_ONLY and Z_RLE compression for images and other     specialized compression   - Added flush options Z_BLOCK to deflate() and Z_TREES to inflate() for     finer control   - Added inflateReset2() and inflateMark() functions, the latter to aid     in random access applications   - Added LFS (Large File Summit) support for 64-bit file offsets and many     other portability improvements
However, note, from the zlib release notes:
If you are using libxml version 2.7.6 or earlier, you will need to update
libxml to version 2.7.7 or later before installing zlib version 1.2.4 or
later. libxml 2.7.6 and earlier made unnecessary assumptions about the
undocumented internal structure of zlib that were changed in zlib 1.2.4
and result in libxml crashing. This was fixed in libxml 2.7.7.

OSX 10.6 has libxml 2.7.8 and is ok, but all current SLC and Ubuntu
distributions still have an older libxml (check with xml2-config --version).

Zlib 1.2.5 is between 10 and 20% faster then version 1.2.3. 
To enable it do:
   ./configure --enable-builtin-zlib


git-svn-id: https://root.cern.ch/svn/root/trunk@39730 27541ba8-7e3a-0410-8455-c3a389f83636