fix https://savannah.cern.ch/bugs/index.php?83910.
In TGraphPainter::PaintGraphErrors and related functions, the caps that are drawn when option "[]" is specified are drawn as a sequence of three separate lines. This results in an ugly concave jag at each corner. The caps are now drawn using TPad::PaintPolyLine, the default miter line joins makes those jags go away.


git-svn-id: https://root.cern.ch/svn/root/trunk@40949 27541ba8-7e3a-0410-8455-c3a389f83636