Allow ROOT to be built with cling, the LLVM-based interpreter, as an interpreter backend. This replaces CINT *only for interpreting code*; CINT is still used for reflection and reflection-based calls. To enable, pass ./configure --with-llvm-config=/location/of/llvm/int/bin --enable-cling to configure.
libCling is now built from within ROOT, using the same flags as e.g. used to build clang.
Now that TCintWithCling is not called TCint anymore, the separate dictionary for TCint can be removed; it's back in G__Meta.
The only configure element needed to find and use LLVM is the llvm-config script which is found from --with-llvm-config.
makelib does not link -lCore -lCint to -lCling.


git-svn-id: https://root.cern.ch/svn/root/trunk@41489 27541ba8-7e3a-0410-8455-c3a389f83636