Extend the processLine interface so that we can use it for other evil purposes like looking up names that come as strings and we can't separate them. This simplifies the implementation of TCling.
There we have interfaces that take strings for function names and parameters.
Those have to be resolved as clang::Decls and it is not possible to use easily
Sema for doing it.

For example we need to be able to get the right overload of clang::FunctionDecl
 for given type name, method name and arguments. processLine will let us create
void wrapper() {(ClassInfo::*f)(arglist) = &ClassInfo::method; } 
and return the parsed clang declaration so we can explore it and get the 
information that is relevant for TCling.


git-svn-id: https://root.cern.ch/svn/root/trunk@41931 27541ba8-7e3a-0410-8455-c3a389f83636